Revert "refact: refacting DPlatformHandle to adap treeland"
This reverts commit e33a3945b29dbd1218bdf2b5dab8b2cc79f7fbcb.
