+++ src/plugins/platform/treeland/dtreelandplatformwindowinterface.h
@@ -9,31 +9,20 @@
 #include "dtreelandplatforminterface.h"
 #include <QObject>
 #include <QtWaylandClient/private/qwaylandwindow_p.h>
-#include "private/dplatformwindowinterface_p.h"
 
-DGUI_BEGIN_NAMESPACE
-
-class PersonalizationWindowContext;
-class DTreeLandPlatformWindowInterface : public QObject, public DPlatformWindowInterface
+DGUI_USE_NAMESPACE
+class DTreeLandPlatformWindowInterface : public QObject
 {
     Q_OBJECT
 public:
-    explicit DTreeLandPlatformWindowInterface(QWindow *window, DPlatformHandle *platformHandle, QObject *parent = nullptr);
-    ~DTreeLandPlatformWindowInterface() override;
-
+    explicit DTreeLandPlatformWindowInterface(QObject *parent = nullptr, QWindow *window = nullptr);
+    ~DTreeLandPlatformWindowInterface();
+    bool setEnabledNoTitlebar(bool enable);
+    void setEnableBlurWindow(bool enable);
+    void doSetEnabledNoTitlebar();
+    void doSetEnabledBlurWindow();
+    [[nodiscard]]QWindow *getWindow() const { return m_window; }
     void initWaylandWindow();
-
-    void setEnabled(bool enabled) override;
-    bool isEnabled() const override;
-
-    bool isEnabledNoTitlebar() const override;
-    bool setEnabledNoTitlebar(bool enable) override;
-
-    int windowRadius() const override;
-    void setWindowRadius(int windowRadius) override;
-
-    bool enableBlurWindow() const override;
-    void setEnableBlurWindow(bool enableBlurWindow) override;
 
 public slots:
     void onSurfaceCreated();
@@ -42,17 +31,14 @@ public slots:
 private:
     PersonalizationWindowContext *getWindowContext();
     void handlePendingTasks();
-    void doSetEnabledNoTitlebar();
-    void doSetWindowRadius();
-    void doSetEnabledBlurWindow();
 
+private:
+    QWindow *m_window = nullptr;
     QQueue<std::function<void()>> m_pendingTasks;
     PersonalizationManager *m_manager = nullptr;
     PersonalizationWindowContext *m_windowContext = nullptr;
-    bool m_isNoTitlebar = false;
+    bool m_isNoTitlebar = true;
     bool m_isWindowBlur = false;
-    int m_radius = 0;
 };
 
-DGUI_END_NAMESPACE
 #endif // DTREELANDPLATFORMWINDOWINTERFACE_H
